requirements.txt
streamlit==1.38.0
pillow==10.4.0
rembg==2.0.59